refactor: correct import paths for OptionsUI post move
