+++ public/app/core/components/OptionsUI/multiSelect.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 
 import { FieldConfigEditorProps, SelectFieldConfigSettings, SelectableValue } from '@grafana/data';
-
-import { MultiSelect } from '../Select/Select';
+import { MultiSelect } from '@grafana/ui';
 
 interface State<T> {
   isLoading: boolean;
