+++ public/app/features/dashboard/components/AnnotationSettings/AnnotationSettingsEdit.tsx
@@ -4,7 +4,8 @@ import { useAsync } from 'react-use';
 import { AnnotationQuery, DataSourceInstanceSettings, getDataSourceRef } from '@grafana/data';
 import { selectors } from '@grafana/e2e-selectors';
 import { DataSourcePicker, getDataSourceSrv } from '@grafana/runtime';
-import { Checkbox, CollapsableSection, ColorValueEditor, Field, HorizontalGroup, Input } from '@grafana/ui';
+import { Checkbox, CollapsableSection, Field, HorizontalGroup, Input } from '@grafana/ui';
+import { ColorValueEditor } from 'app/core/components/OptionsUI/color';
 import StandardAnnotationQueryEditor from 'app/features/annotations/components/StandardAnnotationQueryEditor';
 
 import { DashboardModel } from '../../state/DashboardModel';
